Module classicube_helpers::events::gfx

Structs

ContextLostEvent

ContextLostEventHandler

Context is destroyed after having been previously created */

ContextRecreatedEvent

ContextRecreatedEventHandler

Context is recreated after having been previously lost */

LowVRAMDetectedEvent

LowVRAMDetectedEventHandler

Insufficient VRAM detected, need to free some GPU resources */

ProjectionChangedEvent

ProjectionChangedEventHandler

Projection matrix has changed */

ViewDistanceChangedEvent

ViewDistanceChangedEventHandler

View/fog distance is changed */